+++ cub3d.h
@@ -52,7 +52,15 @@ typedef struct s_wall
 	double	hit;
 	int  start;
 	int  end;
-	char	side;	
+	char	side;
+	double angle;
+	int px_x;
+	int px_y;
+	int x_o;
+	double y_o_step;
+	double start_tex;
+	int y_o;
+	unsigned int tex_y;
 } t_wall;
 
 typedef struct s_playa
@@ -179,6 +187,12 @@ void		free_map(t_textures *textures);
 int			checking_map(t_textures *textures, char *line, int number);
 int			some_rubish_fun(t_textures *textures);
 void	*ft_memcpy(void *dst, const void *src, size_t n);
+
+///drawing wall
+int	norm_color(int c);
+void drawing_ceil_floor(int px_y, int px_x, t_cub *data, t_wall cur );
+mlx_texture_t *get_wall_color(t_wall cur, double angle, t_cub *data);
+
 
 int			flood_fill(char **map, int rows, int cols, int x, int y);
 void		print_map(t_textures *textures);
